refactor: moved payment table to separate component.
